+++ CHANGES.rst
@@ -25,6 +25,7 @@ Changelog
   to the audio data passing through
 - added plugin hierarchy for `WatermarkDetector` plugins and `watermark-detector` filter that applies a watermark
   detection algorithm to the audio data passing through
+- added `adc-phonemize` command-line tool for applying phonemizer to text or text files
 
 
 0.1.0 (2025-10-31)